+++ src/plugins/restoreAssetToCache.ts
@@ -52,9 +52,7 @@ export function restoreAssetToCache(config: RestoreAssetToCacheConfig): Plugin {
             }
 
             try {
-                const headers = new Headers(event.request.headers);
-                headers.set(context.passthroughHeader!, '1');
-                const response = await fetch(new Request(event.request, { headers }));
+                const response = await context.fetchPassthrough!(event.request);
                 if (response.ok) {
                     await cache.put(event.request, response.clone());
                 }